issues mentioned on CM-144 are fixed, The UI is different, there needs to be a comma added to the value
The spelling of balance if wrong (check from Zepplin)

The size of the card is large too. The space between balance and amount should actually be lesser

The UI shows the previous value even after cancellation. The actual amount deducted is same

The fees is Rs. 75, not Rs. 50

The fees should be deducted from the amount entered. Right now it is being added to the amount entered.

The error of lesser balance should take precedence over the error for limits because the limits can be changed

All issues mentioned by M zakir bro are also fixed